Adding nested media object.
Your current method of having comments on a post works fine, but this method also works fine, and doesn't require you to use <ul> tags to get the proper positioning.

As a side-note, when using a <ul> child, the children should always be <li>s, not <div>s. This issue is related to something called Semantic Markup, which basically just says that you should use HTML tags as they are intended (<ul>/<ol> for lists, <table> for tables, <p> for blocks of text, <h1> for page headings, <div> for anything), because things will act more uniformly across browsers and your HTML becomes descriptive of the actual content it is representing.